+++ Python Part/main.py
@@ -6,8 +6,7 @@ Description: multi neural network implementation
 # timeit package for measuring the run speed
 
 import winsound
-import timeit
-import time
+
 from utils import constants
 from utils.read_mnist_data import read_mnist_data
 from network_models import *
@@ -21,7 +20,7 @@ images_test, labels_test = read_mnist_data(constants.MNIST_TEST_FILEPATH, test_s
 
 network_models = network_models(
     images_train=images_train, labels_train=labels_train, images_test=images_test, labels_test=labels_test,
-    n_kernels=3, kernel_size=3, epoch=40, alpha=0.1, hidden=100
+    n_kernels=3,kernel_size=3, epoch=40, alpha=0.1, hidden=100
 )
 
 # network_models.model_one()
@@ -30,13 +29,7 @@ network_models = network_models(
 # network_models.model_six()
 # winsound.Beep(440, 1000)
 
-# network_models.model_one()
-# network_models.model_two()
-# network_models.model_three()
-# network_models.model_four()
-# network_models.model_five()
-network_models.model_six()
-# network_models.model_seven()
+network_models.model_seven()
 winsound.Beep(440, 1000)
 
 winsound.Beep(800, 500)
