+++ Python Part/utils/report_generator.py
@@ -4,8 +4,7 @@ import matplotlib.pyplot as plt
 from utils.date_helper import get_time_format
 
 
-def generate_report(accuracy_train, accuracy_test, epoch_size, training_sample_size, testing_sample_size, alpha, input,
-                    hidden, output, model_name, hidden_weights, output_weights):
+def generate_report(accuracy_train, accuracy_test,epoch_size,training_sample_size,testing_sample_size,alpha,input,hidden,output,model_name):
     print("final training accuracy = %.3f" % accuracy_train[-1])
     print("final test accuracy = %.3f" % accuracy_test[-1])
 
@@ -16,10 +15,10 @@ def generate_report(accuracy_train, accuracy_test, epoch_size, training_sample_s
     plt.xlabel('Epoch')
     plt.ylabel('Accuracy')
 
-    plot.savefig("temp/results/figures/{}_{}.png".format(model_name, get_time_format()))
+    plot.savefig("temp/results/figures/{}_{}.png".format(model_name,get_time_format()))
 
     # Open a file for writing the results
-    with open('temp/results/text/{}_{}.csv'.format(model_name, get_time_format()), 'w') as f:
+    with open('temp/results/text/{}_{}.csv'.format(model_name,get_time_format()), 'w') as f:
         # Write the information about the experiment
         f.write(f"Model Name: {model_name}\n")
         f.write(f"Epoch Size: {epoch_size},")
@@ -27,8 +26,7 @@ def generate_report(accuracy_train, accuracy_test, epoch_size, training_sample_s
         f.write(f"Testing Sample Size: {testing_sample_size},")
         f.write(f"Alpha: {alpha}\n")
         f.write(f"Hyper Parameters - Input: {input}, Hidden: {hidden}, Output: {output}\n ")
-        f.write(f"hidden weights ${hidden_weights}\n")
-        f.write(f"output weights ${output_weights}\n")
+
         # Write the header for the accuracy results
         f.write("Epoch,Train Accuracy,Test Accuracy\n")
 
@@ -37,4 +35,6 @@ def generate_report(accuracy_train, accuracy_test, epoch_size, training_sample_s
             f.write(f"{i + 1},{acc_train},{acc_test}\n")
 
 
-
+        # TODO: pass the weights to the function after the end of the training, so that it get saved in the file
+        # f.write(f"hidden weights ${hidden_weights}\n")
+        # f.write(f"output weights ${output_weights}\n")
